[+] Factory function for session login
